+++ lib/spack/spack/test/pytest.ini
@@ -1,7 +1,7 @@
 # content of pytest.ini
 [pytest]
 addopts = --durations=20 -ra
-testpaths = lib/spack/spack/test
+testpaths = .
 python_files = *.py
 markers =
   db: tests that require creating a DB
